+++ examples/nuxt/app.vue
@@ -1,0 +1,7 @@
+<template>
+  <main>  
+    <NuxtLayout name="default">
+      <NuxtPage/>
+    </NuxtLayout>
+  </main>
+</template>