+++ sandbox/CHANGELOG.md
@@ -26,4 +26,4 @@
 ### Patch Changes
 
 - Updated dependencies [02f38a1]
-  - @pglite-utils/prisma-adapter@0.1.0
+  - pglite-prisma-adapter@0.1.0